pyproject.toml
[tool.poetry]
name = "discord-link-bot"
version = "1.0.0"
description = "A Discord bot that links users with laravel application using tokens from a MySQL database."
authors = ["Arthuurlijk"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
discord-py = "^2.2.2"
fastapi = "^0.103.1"
mysql-connector-python = "^8.1.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"